requirements.txt
requests==2.31.0
atlassian-python-api==3.41.3
markdown==3.5.1
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7
cryptography==41.0.5
beautifulsoup4==4.12.2
html2text==2020.1.16
rich==13.6.0 